+++ Models/user.cs
@@ -37,7 +37,7 @@ public class user
     public int flag { get; set; } = 0;
 
     public string? image { get; set; }
-    public int? coins { get; set; }
+    public double? coins { get; set; }
 
     public ICollection<userRole> userRoles { get; } = new List<userRole>();
